Run tests only on primary targets. (#13184)
Distinguish between "target" (to be compiled) and "primary_target" (to
be tested/analyzed/etc.) in move-model/move-compiler-v2, so that
tests/etc. can be run on just the primary source inputs.  Primary
targets are distinguished by keeping sources separate on their way to
the move model through various interfaces.  In move-model internals
(model.rs), the `is_dep` property is negated as `is_target` (to be
compiled), while the root packages (to be tested or whatever) are
identified as `is_prime_target`.

GitOrigin-RevId: 5422dc17a873fa13faa15762cb7a7e22a5372ad5