[move] View APIs instead of Move resolvers (#13233)
Previously, `MoveResolver`s have been used to access storage bytes by
the VM, Move APIs, indexer, etc. As a result, this created an unnecessary
coupling between the VM and the rest of the system, making code
maintainability harder, error-prone and hard to evolve (e.g., implementing
a new loader).

This refactors code so that a separate trait is used by indexers/APIs.

GitOrigin-RevId: 9806be87a913124363b62721ff9a9e04d2c5c029