Fix abigen test by fixing dep path to avoid catching built files if present (#13380)
Fix dependency path in a move-abigen test so it still works with a built
move-stdlib, while adding comments and
some Debug implementations to make it easier to sort out what paths
are present where in the code.

This PR also adds somewhat unnecessary dependences between stdlib
builds and tests using them, not because they use the built libs (they
should be reading the sources, not the built packages), but because
this leads to more deterministic output when paths are not specified
correctly.

GitOrigin-RevId: d57183418ba38281a144e62c1fe6752ef07e9727