[compiler-v2] Bytecode verification failure error message (#13469)
* [compiler-v2] Bytecode verification failure error message

This close #12709 and mitigates #12817.

If we had a bytecode verification failure post file format verification, until now we gave a cryptical error. We also have some weird corner cases where the verifier fails while reference safety is clearly maintained.

This improves the error message on bv failure, as well as prints out a special tailored message for the corner cases, which we will be hopefully able to fix with a new bytecode verifier. The full VMError detail
is omitted unless the env var MVC_BACKTRACE is set to true.

* Addressing reviewer comments

GitOrigin-RevId: 1d9cd70677a2fa76e993d5ad5a281d046dec6a19