[compiler-v2] Enum types frontend (#13577)
* [compiler-v2] Enum types frontend

This implements the frontend (parser, context checker, and model representation) of struct variants aka 'enum types', including the `enum` type declaration, as well as the `match` expression.

The internal design principle is that enum types are not a new kind of types, but just a variation of the existing struct types, which now have multiple variants. This allows the implementation to have a relative small footprint. The same principle is likely later to be applied for the VM implementation.

Many of the existing code for dealing with struct pack and unpack can be reused for enum types. Some changes in the expansion phase and the model builder have to be applied, as a struct name can now be accompanied by an optional variant name. The model builder has also been slightly refactored and cleaned up to deal with the new features.

The implementation does not yet check for exhaustive matches and similar conditions. Those checks are naturally done later by the code which compiles matches to bytecode.

There are some general changes to baseline files unrelated to the functionality of this PR but resulting from refining the Model dump representation, as well as minor bug fixes. For the tests relevant to this PR, see `test/checking/variants/`.

* Check for correct naming of variants, and resolve unqualified variants in match

* addressing reviewer comments. Also adding common field resolution for variants.

GitOrigin-RevId: 97a555b0d13faa1f498be4debeafd646664e9e45