[compiler-v2] Enable empty field list (#14415)
* [compiler-v2] Enable empty field list, remove dummy field from positional

Enables struct declarations of the form `struct T;`, which is interpreted as `struct T();`. This makes the declaration of type tags more concise.

Also omits generation of the 'dummy field' for positional structs. As they are not in the v1 language, we can skip this compatibility issue.

* Reverting change to remove dummy field, unfortunately it is needed and the VM checks this.

Adding more tests.

GitOrigin-RevId: cf7e840b7eb0844c5f0ddaeaa85b825e607e465e