[compiler-v2][decompiler] Pulling fat-loop algorithm out of the prover (#14488)
The prover actually has a full-fledged loop analysis for stackless bytecode for loop invariant instrumentation and unrolling. This loop analysis is a major part of the logic for decompilation. This PR pulls out this functionality in a new general purpose module, `fat_loop`. Its mostly a refactoring of existing code.

GitOrigin-RevId: 51f896c2c71f98c0a5e3e8203ea7b8eefe5477ca